requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2.4
django-autocomplete-light==3.12.1
django-jazzmin==3.0.1
django-phonenumber-field==8.0.0
fonttools==4.58.4
phonenumbers==9.0.2
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
